Correctifs 9/10 : WCAG, dark mode, YouTube, reduced-motion, favicon - 06 Sable & Bronze : fond inverse creme dominant, texte vert fonce (#1A4A3A) pour WCAG AA - 05 Neon Nights : fix syntaxe class"section", rose->#FF69B4 pour contraste - 03 Scene Vintage : dark mode CSS, scroll indicator cliquable, multi-favicon - YouTube embeds : placeholders "Video a venir" dans les 6 propositions - prefers-reduced-motion : ajoute dans les 6 animations

+++ scripts/render_animation.py
@@ -97,7 +97,7 @@ def make_frame(t):
     outer_r = 165 * SCALE
     timbre_prog = min(1, max(0, (t - 0.6) / 0.6))
     if timbre_prog > 0:
-        for ang in range(0, 360 * timbre_prog, 8):
+        for ang in range(0, int(360 * timbre_prog), 8):
             a1 = math.radians(ang)
             a2 = math.radians(min(ang + 4, 360 * timbre_prog))
             draw.arc([CX - outer_r, CY - outer_r, CX + outer_r, CY + outer_r],
